pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "devtools_scicomp_project_2025"
version = "0.1"
dependencies = ["chex",
                "dm-haiku",
                "flax",
                "jax",
                "numpy"]
requires-python = ">=3.11"
authors = [{name = "Stefano Campanella"}]
description = 'Assignment for the course "Development Tools for Scientific Computing 2024/5" held at SISSA'
readme = {file = "README.md", content-type = "text/markdown"}
license = "MIT"

[project.optional-dependencies]
interactive = ["jupyterlab", "ipywidgets", "matplotlib"]
profile = ["tensorboard_plugin_profile"]
train = ["torch", "torchvision", "optax", "tensorboard"]
test = ["pytest", "pytest-cov"]


[project.urls]
Homepage = "https://github.com/stefanocampanella/devtools_scicomp_project_2025"

[tool.setuptools]
packages = ["causal_conv_arm"]